Fixes #11: Price field radio buttons display amount regardless of configuration.
Based on a read of the code, checkboxes were also affected by this bug. The fundamental problem was that the widget
templates didn't take the displayPrice property into account. For consistency, I updated the normalizer to use the
new normalizeBoolean() method, reducing doubt on the client side about the return format.